General SubmodelComponents can now be saved to a json format
Format for SimpleRectangularSubmodelComponents has changed.
Format for SubmodelComponents holds information on Pin positions.
Both SubmodelComponents and SimpleRect... can be created through the
GUICustomComponentCreator (SubmodelComponents are created as
SimpleRect..., as SubmodelComponent is abstract). This can be expanded
upon for the addition of more subclasses to SubmodelComponent. Included
a class JsonExample for testing a few standard cases.